+++ services/familyservice/family_service.go
@@ -73,7 +73,7 @@ type FamilyService interface {
 	// 解散家族
 	DissolutionFamily(ctx context.Context, userID uint64, familyID int32) error
 	//设置家族群组id
-	SetFamilyGroupID(ctx context.Context, familyID int32, groupID int32) error
+	SetFamilyGroupID(ctx context.Context, familyID int32, groupID int64) error
 }
 
 var GlobalFamilyService FamilyService
